feat(meetings): §2 Track D — dedicated meeting detail page (minutes-first)
The meeting deliverable (minutes: summary/decisions/action-items) was buried
inside an expanded list card below the full transcript, so generated drafts sat
unconfirmed and invisible. Track D moves it to a dedicated `/meetings/{id}` page
where the minutes ARE the default view and the raw transcript is secondary +
collapsed, with a prominent draft-confirm nudge so drafts don't rot.

- New `pages/MeetingDetailPage.tsx` at `/meetings/:id` (gated on
  meeting_transcription_enabled; minutes-first view gated on
  meeting_minutes_enabled). `useMeeting(id)` → GET /api/meetings/{id}.
- Completed list cards now LINK to the detail page (inline expand removed — the
  card affordance was the Phase-0 first slice); the draft-ready badge links there.
- Extract shared building blocks to `components/meetings/`
  (StatusBadge/TranscriptView/MinutesPanel/ProjectSelect), consumed by both pages.
- de/en i18n (backToList, notFound, stillProcessing, transcriptSection,
  openDetail(For), draftNudgeTitle/Body). Minutes mutations now also invalidate
  the detail query so the draft banner tracks generate/confirm.
- Tests: rewrite MeetingsPage.test.tsx for link-only cards; new
  MeetingDetailPage.test.tsx (minutes-first, draft nudge, transcript toggle,
  not-found, delete→navigate, generate/edit/confirm, relabel).
