+++ src/frontend/src/i18n/locales/de.json
@@ -727,7 +727,15 @@
       "failedToSave": "Protokoll konnte nicht gespeichert werden",
       "failedToConfirm": "Protokoll konnte nicht bestätigt werden",
       "failedToDelete": "Protokoll konnte nicht verworfen werden"
-    }
+    },
+    "backToList": "Zurück zur Übersicht",
+    "notFound": "Besprechung nicht gefunden",
+    "stillProcessing": "Die Besprechung wird noch verarbeitet — Transkript und Protokoll erscheinen, sobald sie fertig ist.",
+    "transcriptSection": "Transkript anzeigen",
+    "openDetail": "Details öffnen",
+    "draftNudgeTitle": "Protokoll-Entwurf wartet auf Prüfung",
+    "draftNudgeBody": "Ein Entwurf wurde erzeugt, aber noch nicht bestätigt. Prüfe ihn unten und bestätige ihn, damit er ins Dokument übernommen wird.",
+    "openDetailFor": "Details für {{title}} öffnen"
   },
   "tasks": {
     "title": "Aufgaben",
